+++ src/facets/MarketplaceFacet.sol
@@ -12,15 +12,15 @@ contract MarketplaceFacet is IMarketplace {
     //                             EXTERNAL FUNCTIONS
     //////////////////////////////////////////////////////////////////////////*//
 
-    function mintTicket(uint56 _ticketId, FeeType _feeType, address _buyer) external payable returns (uint40) {
+    function mintTicket(uint64 _ticketId, FeeType _feeType, address _buyer) external payable returns (uint40) {
         return _ticketId._mintTicket(_feeType, _buyer);
     }
 
-    function setTicketFees(uint56 _ticketId, FeeType[] calldata _feeTypes, uint256[] calldata _fees) external {
+    function setTicketFees(uint64 _ticketId, FeeType[] calldata _feeTypes, uint256[] calldata _fees) external {
         _ticketId._setTicketFees(_feeTypes, _fees);
     }
 
-    function withdrawTicketBalance(uint56 _ticketId, FeeType _feeType, address _to) external {
+    function withdrawTicketBalance(uint64 _ticketId, FeeType _feeType, address _to) external {
         _ticketId._withdrawTicketBalance(_feeType, _to);
     }
 
@@ -32,7 +32,7 @@ contract MarketplaceFacet is IMarketplace {
     //                               VIEW FUNCTIONS
     //////////////////////////////////////////////////////////////////////////*//
 
-    function isFeeEnabled(uint56 _ticketId, FeeType _feeType) external view returns (bool) {
+    function isFeeEnabled(uint64 _ticketId, FeeType _feeType) external view returns (bool) {
         return _ticketId._isFeeEnabled(_feeType);
     }
 
@@ -40,11 +40,11 @@ contract MarketplaceFacet is IMarketplace {
         return _feeType._getFeeTokenAddress();
     }
 
-    function getTicketFee(uint56 _ticketId, FeeType _feeType) external view returns (uint256) {
+    function getTicketFee(uint64 _ticketId, FeeType _feeType) external view returns (uint256) {
         return _ticketId._getTicketFee(_feeType);
     }
 
-    function getAllFees(uint56 _ticketId, FeeType _feeType)
+    function getAllFees(uint64 _ticketId, FeeType _feeType)
         external
         view
         returns (uint256 ticketFee_, uint256 hostItFee_, uint256 totalFee_)
@@ -52,7 +52,7 @@ contract MarketplaceFacet is IMarketplace {
         return _ticketId._getFees(_feeType);
     }
 
-    function getTicketBalance(uint56 _ticketId, FeeType _feeType) external view returns (uint256) {
+    function getTicketBalance(uint64 _ticketId, FeeType _feeType) external view returns (uint256) {
         return _ticketId._getTicketBalance(_feeType);
     }
