+++ src/interfaces/IFactory.sol
@@ -12,21 +12,21 @@ interface IFactory {
     function createTicket(TicketData calldata _ticketData, FeeType[] calldata _feeTypes, uint256[] calldata _fees)
         external;
 
-    function updateTicket(TicketData calldata _ticketData, uint56 _ticketId) external;
+    function updateTicket(TicketData calldata _ticketData, uint64 _ticketId) external;
 
     //*//////////////////////////////////////////////////////////////////////////
     //                               VIEW FUNCTIONS
     //////////////////////////////////////////////////////////////////////////*//
 
-    function ticketCount() external view returns (uint56);
+    function ticketCount() external view returns (uint64);
 
-    function ticketExists(uint56 _ticketId) external view returns (bool);
+    function ticketExists(uint64 _ticketId) external view returns (bool);
 
-    function ticketData(uint56 _ticketId) external view returns (FullTicketData memory);
+    function ticketData(uint64 _ticketId) external view returns (FullTicketData memory);
 
     function allTicketData() external view returns (FullTicketData[] memory);
 
-    function adminTickets(address _ticketAdmin) external view returns (uint56[] memory);
+    function adminTickets(address _ticketAdmin) external view returns (uint64[] memory);
 
     function adminTicketData(address _ticketAdmin) external view returns (FullTicketData[] memory);
 
@@ -36,9 +36,9 @@ interface IFactory {
 
     function hostItTicketHash() external pure returns (bytes32);
 
-    function ticketHash(uint56 _ticketId) external pure returns (bytes32);
+    function ticketHash(uint64 _ticketId) external pure returns (bytes32);
 
-    function mainAdminRole(uint56 _ticketId) external pure returns (uint256);
+    function mainAdminRole(uint64 _ticketId) external pure returns (uint256);
 
-    function ticketAdminRole(uint56 _ticketId) external pure returns (uint256);
+    function ticketAdminRole(uint64 _ticketId) external pure returns (uint256);
 }
