+++ src/interfaces/IMarketplace.sol
@@ -8,11 +8,11 @@ interface IMarketplace {
     //                             EXTERNAL FUNCTIONS
     //////////////////////////////////////////////////////////////////////////*//
 
-    function mintTicket(uint56 ticketId, FeeType feeType, address buyer) external payable returns (uint40);
+    function mintTicket(uint64 ticketId, FeeType feeType, address buyer) external payable returns (uint40);
 
-    function setTicketFees(uint56 ticketId, FeeType[] calldata feeTypes, uint256[] calldata fees) external;
+    function setTicketFees(uint64 ticketId, FeeType[] calldata feeTypes, uint256[] calldata fees) external;
 
-    function withdrawTicketBalance(uint56 ticketId, FeeType feeType, address to) external;
+    function withdrawTicketBalance(uint64 ticketId, FeeType feeType, address to) external;
 
     function withdrawHostItBalance(FeeType feeType, address to) external;
 
@@ -20,17 +20,17 @@ interface IMarketplace {
     //                               VIEW FUNCTIONS
     //////////////////////////////////////////////////////////////////////////*//
 
-    function isFeeEnabled(uint56 ticketId, FeeType feeType) external view returns (bool);
+    function isFeeEnabled(uint64 ticketId, FeeType feeType) external view returns (bool);
 
     function getFeeTokenAddress(FeeType feeType) external view returns (address);
 
-    function getTicketFee(uint56 ticketId, FeeType feeType) external view returns (uint256);
-    function getAllFees(uint56 ticketId, FeeType feeType)
+    function getTicketFee(uint64 ticketId, FeeType feeType) external view returns (uint256);
+    function getAllFees(uint64 ticketId, FeeType feeType)
         external
         view
         returns (uint256 ticketFee, uint256 hostItFee, uint256 totalFee);
 
-    function getTicketBalance(uint56 ticketId, FeeType feeType) external view returns (uint256);
+    function getTicketBalance(uint64 ticketId, FeeType feeType) external view returns (uint256);
 
     function getHostItBalance(FeeType feeType) external view returns (uint256);
